+++ includes/api/mybooking-plugin-activities-api-client.php
@@ -7,7 +7,9 @@
 
     const GET_ACTIVITIES = '/api/booking-activities/frontend/activities';
     const GET_ACTIVITY = '/api/booking-activities/frontend/activities/';
-    
+    const GET_DESTINATIONS = '/api/booking-activities/frontend/destinations';
+    const GET_CATEGORIES = '/api/booking-activities/frontend/categories';
+
     private $url_prefix;
     private $api_key;
 		
@@ -19,18 +21,91 @@
 	  }
 
     /**
+     * Get the destinations
+     */
+    public function get_destinations() {
+
+          $url = $this->url_prefix.MyBookingApiClient::GET_DESTINATIONS;
+
+          if ( isset($this->api_key) && !empty($this->api_key)) {
+            $url = $url.'&api_key='.$this->api_key;
+          }
+
+          // Query data
+          $request = wp_remote_get( $url );
+          if( is_wp_error( $request ) ) {
+            return null; // TODO
+          }
+
+          // Check status code
+          $status_code = wp_remote_retrieve_response_code( $request );
+          if ( $status_code == 404 || $status_code == 500) {
+            return null;
+          }
+
+          // Get the response body
+          $body = wp_remote_retrieve_body( $request );
+          $data = json_decode( $body );
+          return $data; 
+
+    }
+
+    /**
+     * Get the destinations
+     */
+    public function get_categories() {
+
+          $url = $this->url_prefix.MyBookingApiClient::GET_CATEGORIES;
+
+          if ( isset($this->api_key) && !empty($this->api_key)) {
+            $url = $url.'&api_key='.$this->api_key;
+          }
+
+          // Query data
+          $request = wp_remote_get( $url );
+          if( is_wp_error( $request ) ) {
+            return null; // TODO
+          }
+
+          // Check status code
+          $status_code = wp_remote_retrieve_response_code( $request );
+          if ( $status_code == 404 || $status_code == 500) {
+            return null;
+          }
+
+          // Get the response body
+          $body = wp_remote_retrieve_body( $request );
+          $data = json_decode( $body );
+          return $data; 
+
+    }
+
+    /**
      * Get the activities
      *
+     * @param number $destination_id The destination
+     * @param number $family_id The family
+     * @param string $q The query
      * @param number $offset The offset on pagination
      * @param number $limit The number of records on pagination
      */
-    public function get_activities($q, $offset=0, $limit=20) {
+    public function get_activities($destination_id=null, $family_id=null, $q='', $offset=0, $limit=20) {
 
           $url = $this->url_prefix.MyBookingActivitiesApiClient::GET_ACTIVITIES.'?offset='.$offset.'&limit='.$limit;
+          
+          // Destination
+          if ( $destination_id != null && !empty($destination_id) ) {
+            $url.='&destination_id='.$destination_id;
+          }
+          // Family
+          if ( $family_id != null && !empty($family_id) ) {
+            $url.='&family_id='.$family_id;
+          }
+          // Query
           if ( $q != null && !empty($q) ) {
             $url.='&q='.urlencode( $q );
           }
-
+          // Api Key
           if ( isset($this->api_key) && !empty($this->api_key)) {
             $url = $url.'&api_key='.$this->api_key;
           }